Move DeconzEvent to init
